add author controls.

+++ src/main/java/com/tahayvz/publisherapp/controllers/PublishingHouseController.java
@@ -1,8 +1,11 @@
 package com.tahayvz.publisherapp.controllers;
 
 import com.tahayvz.publisherapp.commands.PublishingHouseCommand;
+import com.tahayvz.publisherapp.domain.User;
 import com.tahayvz.publisherapp.services.PublishingHouseService;
+import com.tahayvz.publisherapp.services.UserService;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
@@ -15,41 +18,60 @@ import javax.validation.Valid;
 
 @Slf4j
 @Controller
-public class PublishingHouseController{
+public class PublishingHouseController extends CommonMethods{
 
-    private static final String PUBLISHINGHOUSE_PUBLISHINGHOUSEFORM_URL = "publishinghouse/publishinghouseform";
+    @Autowired
+    private UserService userService;
+
+    private static final String PUBLISHINGHOUSE_PUBLISHINGHOUSEFORM_URL = "publishinghouse/publishinghouseForm";
     private final PublishingHouseService publishingHouseService;
 
     public PublishingHouseController(PublishingHouseService publishingHouseService) {
         this.publishingHouseService = publishingHouseService;
     }
-    
+
     @GetMapping("/publishinghouse")
     public String getIndexPage(Model model){
-        model.addAttribute("publishingHouses", publishingHouseService.getPublishingHouses());
-
-        return "index";
+        User activeUser = userService.findByEmail(getActiveLoggedUserEmail());
+        if(activeUser!=null) {
+            model.addAttribute("publishingHouses", publishingHouseService.getPublishingHouses());
+            return "index";
+        } else{
+            return "redirect:/";
+        }
     }
 
     @GetMapping("/publishinghouse/{id}/show")
     public String showById(@PathVariable String id, Model model){
-
-        model.addAttribute("publishingHouse", publishingHouseService.findById(new Long(id)));
-
-        return "publishinghouse/show";
+        User activeUser = userService.findByEmail(getActiveLoggedUserEmail());
+        if(activeUser!=null) {
+            model.addAttribute("publishingHouse", publishingHouseService.findById(new Long(id)));
+            return "publishinghouse/show";
+        } else{
+            return "redirect:/";
+        }
     }
 
     @GetMapping("publishinghouse/new")
-    public String newPublishinghouse(Model model){
-        model.addAttribute("publishingHouse", new PublishingHouseCommand());
-
-        return "publishinghouse/publishingHouseForm";
+    public String newPublishingHouse(Model model){
+        User activeUser = userService.findByEmail(getActiveLoggedUserEmail());
+        if(activeUser!=null) {
+            model.addAttribute("publishingHouse", new PublishingHouseCommand());
+            return PUBLISHINGHOUSE_PUBLISHINGHOUSEFORM_URL;
+        }else{
+            return "redirect:/";
+        }
     }
 
     @GetMapping("publishinghouse/{id}/update")
     public String updateRecipe(@PathVariable String id, Model model){
-        model.addAttribute("publishingHouse", publishingHouseService.findCommandById(Long.valueOf(id)));
-        return PUBLISHINGHOUSE_PUBLISHINGHOUSEFORM_URL;
+        User activeUser = userService.findByEmail(getActiveLoggedUserEmail());
+        if(activeUser!=null) {
+            model.addAttribute("publishingHouse", publishingHouseService.findCommandById(Long.valueOf(id)));
+            return PUBLISHINGHOUSE_PUBLISHINGHOUSEFORM_URL;
+        } else{
+            return "redirect:/";
+        }
     }
 
     @PostMapping("publishinghouse")
@@ -68,14 +90,19 @@ public class PublishingHouseController{
 
         return "redirect:/publishinghouse/" + savedCommand.getId() + "/show";
     }
-    
+
     @GetMapping("publishinghouse/{id}/delete")
     public String deleteById(@PathVariable String id){
+        User activeUser = userService.findByEmail(getActiveLoggedUserEmail());
+        if(activeUser!=null) {
+            log.debug("Deleting id: " + id);
 
-        log.debug("Deleting id: " + id);
+            publishingHouseService.deleteById(Long.valueOf(id));
+            return "redirect:/index";
+        } else{
+            return "redirect:/";
+        }
 
-        publishingHouseService.deleteById(Long.valueOf(id));
-        return "redirect:/";
     }
 
 }
